+++ examples/visualization/plot_clickable_image.py
@@ -60,6 +60,7 @@ events = np.hstack([events, np.zeros([nepochs, 2])])
 data = np.random.randn(nepochs, nchans, sr * nsec)
 info = mne.create_info(nchans, sr, ch_types='eeg')
 epochs = mne.EpochsArray(data, info, events)
+evoked = epochs.average()
 
 # Using the native plot_topo function with the image plotted in the background
-f = mne.viz.plot_topo(epochs.average(), layout=lt, fig_background=im)
+f = evoked.plot_topo(layout=lt, fig_background=im)