+++ examples/visualization/plot_topo_compare_conditions.py
@@ -21,7 +21,7 @@ import matplotlib.pyplot as plt
 import mne
 
 from mne.io import Raw
-from mne.viz import plot_topo
+from mne.viz import plot_evoked_topo
 from mne.datasets import sample
 
 print(__doc__)
@@ -66,7 +66,7 @@ evokeds = [epochs[name].average() for name in ('left', 'right')]
 colors = 'yellow', 'green'
 title = 'MNE sample data - left vs right (A/V combined)'
 
-plot_topo(evokeds, color=colors, title=title)
+plot_evoked_topo(evokeds, color=colors, title=title)
 
 conditions = [e.comment for e in evokeds]
 for cond, col, pos in zip(conditions, colors, (0.025, 0.07)):