+++ mne/viz/__init__.py
@@ -13,7 +13,7 @@ from ._3d import (plot_sparse_source_estimates, plot_source_estimates,
 from .misc import (plot_cov, plot_bem, plot_events, plot_source_spectrogram,
                    _get_presser, plot_dipole_amplitudes)
 from .evoked import (plot_evoked, plot_evoked_image, plot_evoked_white,
-                     plot_snr_estimate)
+                     plot_snr_estimate, plot_evoked_topo)
 from .circle import plot_connectivity_circle, circular_layout
 from .epochs import (plot_image_epochs, plot_drop_log, plot_epochs,
                      plot_epochs_trellis, _drop_log_stats, plot_epochs_psd)