+++ doc/whats_new.rst
@@ -39,6 +39,8 @@ Changelog
 
     - Add new method :class:`mne.preprocessing.Xdawn` for denoising and decoding of ERP/ERF by `Alexandre Barachant`_
 
+    - Add ``plot_topo`` method to ``Evoked`` object by `Jaakko Leppakangas`_
+
     - Add support for plotting patterns/filters in :class:`mne.decoding.csp.CSP` and :class:`mne.decoding.base.LinearModel` by `Romain Trachel`_
 
     - Add new object :class:`mne.decoding.base.LinearModel` for decoding M/EEG data and interpreting coefficients of linear models with patterns attribute by `Romain Trachel`_ and `Alex Gramfort`_
@@ -49,8 +51,9 @@ Changelog
 
     - Deprecated `lws` and renamed `ledoit_wolf` for the ``reg`` argument in :class:`mne.decoding.csp.CSP` and :class:`mne.preprocessing.Xdawn` by `Romain Trachel`_ 
 
-
     - Add interactive plotting of topomap from time-frequency representation by `Jaakko Leppakangas`_
+
+    - Add ``plot_topo`` method to ``Evoked`` object by `Jaakko Leppakangas`_
 
 BUG
 ~~~
@@ -64,6 +67,11 @@ BUG
     - Fix picking channels from forward operator could return a channel ordering different from ``info['chs']`` by `Chris Bailey`_
 
     - Fix dropping of events after downsampling stim channels by `Marijn van Vliet`
+
+API
+~~~
+
+    - Rename and deprecate ``mne.viz.plot_topo`` for ``mne.viz.plot_evoked_topo`` by `Jaakko Leppakangas`_
 
 .. _changes_0_9: